# docs/interfaces/schemas/p2-events.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "P2 Core event and UI contract proposal",
    "version": "1.0",
    "description": "Core event schemas; the task-events route is implemented in PR2."
  },
  "paths": {
    "/api/v1/tasks/{task_id}/events": {
      "get": {
        "operationId": "listCoreTaskEvents",
        "x-implementation-status": "implemented-pr2",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "APIResponse containing task events in sequence order",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BehaviorEvent"
                      }
                    },
                    "error": {
                      "type": [
                        "object",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GatewayDecisionType": {
        "enum": [
          "ALLOW",
          "DENY",
          "REQUIRE_CONFIRMATION",
          "REQUIRE_REPLAN"
        ],
        "title": "GatewayDecisionType",
        "type": "string"
      },
      "GatewayReasonCode": {
        "enum": [
          "ALLOWED",
          "USER_DENY",
          "OUT_OF_CONTRACT",
          "SKILL_LIMIT",
          "SYSTEM_DENY",
          "RESOURCE_MISMATCH",
          "VERSION_STALE",
          "CONFIRMATION_REQUIRED",
          "CHECK_UNAVAILABLE",
          "ADAPTER_BYPASS",
          "SIGNATURE_INVALID",
          "INTENT_DEVIATION",
          "INTENT_STOPPED"
        ],
        "title": "GatewayReasonCode",
        "type": "string"
      },
      "IntentAssessment": {
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "default": "intent-rule-v1",
            "title": "Schema Version",
            "type": "string"
          },
          "risk_score": {
            "maximum": 1,
            "minimum": 0,
            "title": "Risk Score",
            "type": "number"
          },
          "trigger_dimensions": {
            "items": {
              "type": "string"
            },
            "title": "Trigger Dimensions",
            "type": "array"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "string"
          },
          "detector_version": {
            "default": "bounded-report-rule:1",
            "title": "Detector Version",
            "type": "string"
          },
          "contract_version": {
            "title": "Contract Version",
            "type": "integer"
          },
          "contract_digest": {
            "title": "Contract Digest",
            "type": "string"
          },
          "step_index": {
            "title": "Step Index",
            "type": "integer"
          },
          "disposition": {
            "title": "Disposition",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "risk_score",
          "policy_version",
          "contract_version",
          "contract_digest",
          "step_index",
          "disposition",
          "reason",
          "enabled"
        ],
        "title": "IntentAssessment",
        "type": "object"
      },
      "BehaviorEvent": {
        "additionalProperties": false,
        "description": "A stored event with all twelve plan fields explicitly present.\n\nNullable facts are explicit; consumers must not invent missing provenance.\nThe store allocates sequence before validating this persisted representation.",
        "properties": {
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "event_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Event Id",
            "type": "string"
          },
          "sequence": {
            "maximum": 9007199254740991,
            "minimum": 1,
            "title": "Sequence",
            "type": "integer"
          },
          "task_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Task Id",
            "type": "string"
          },
          "parent_event_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "\\S",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Event Id"
          },
          "type": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Type",
            "type": "string"
          },
          "actor": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Actor",
            "type": "string"
          },
          "source_ref": {
            "maxLength": 512,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Source Ref",
            "type": "string"
          },
          "object_digest": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Digest"
          },
          "state": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "\\S",
            "title": "State",
            "type": "string"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GatewayDecisionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "result_digest": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Digest"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "\\S",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "reason_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GatewayReasonCode"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "evidence_refs": {
            "items": {
              "maxLength": 512,
              "minLength": 1,
              "pattern": "\\S",
              "type": "string"
            },
            "maxItems": 100,
            "title": "Evidence Refs",
            "type": "array"
          },
          "intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IntentAssessment"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "prepared_effect_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "\\S",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Prepared Effect Id"
          },
          "effect_descriptor_digest": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Effect Descriptor Digest"
          },
          "result_commitment": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "pattern": "\\S",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Result Commitment"
          },
          "commit_epoch": {
            "anyOf": [
              {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Commit Epoch"
          },
          "execution_receipt_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "\\S",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Execution Receipt Id"
          },
          "compliance_proof_ref": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "pattern": "\\S",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Compliance Proof Ref"
          }
        },
        "required": [
          "event_id",
          "sequence",
          "task_id",
          "parent_event_id",
          "type",
          "actor",
          "source_ref",
          "object_digest",
          "state",
          "decision",
          "result_digest",
          "occurred_at"
        ],
        "title": "BehaviorEvent",
        "type": "object"
      },
...
    }
  }
}